requirements.txt
torch
torchaudio
pyannote.audio
moviepy
librosa
soundfile
tqdm
faster-whisper
